requirements.txt
python-whois==0.9.6
requests==2.32.5
python-dateutil==2.9.0.post0
